divine/nostr-kafka-bridge/requirements.txt
websockets>=12.0
kafka-python>=1.4.7
